requirements.txt
flask
pdfminer.six
scikit-learn
numpy
pandas
nltk
reportlab
